feat(sql): add a valueMapper
The `valueMapper` works similarly to the `fieldMapper` but acts on values. It comes with a set of
default helper functions intended to be used in LIKE scenarios (`~`) where you commonly want `foo ~
bar` to translate into `foo LIKE "%bar"`.

The helpers also escapes characters that could lead to a SQL injection.