chore(docs): improve user-facing documentation
A rewrite of all README's to optimize for the different use-cases as well as
scoping it for each package. Rewrote it based on feedback, it is clear that
I was biased while writing these.

Also, the api was out of date due to maintaining it by hand - need to
figure out how to automate this.

PR: https://github.com/jbergstroem/filtron/pull/98